requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
flask-swagger-ui==4.11.1
